+++ client/stories/hooks/useAutoSequence.ts
